adding portfolio selection to main page (wip)
