+++ backend/node_app/modules/eda/edaSearchHandler.js
@@ -134,7 +134,7 @@ class EdaSearchHandler extends SearchHandler {
 
 	async documentSearch(req, body, clientObj, userId) {
 		try {
-			const permissions = req.permissions ? req.permissions : [];
+			const permissions = req.permissions ? req.permissions.map((permission) => permission.toLowerCase()) : [];
 			const { getIdList, selectedDocuments, expansionDict = {}, forGraphCache = false, forStats = false } = body;
 			const [parsedQuery, searchTerms] = this.searchUtility.getEsSearchTerms(body);
 			body.searchTerms = searchTerms;
@@ -142,7 +142,12 @@ class EdaSearchHandler extends SearchHandler {
 
 			const { esClientName, esIndex } = clientObj;
 			let esQuery = '';
-			if (permissions.includes('View EDA') || permissions.includes('eda Admin')) {
+			console.log(permissions);
+			if (
+				permissions.includes('view eda') ||
+				permissions.includes('eda admin') ||
+				permissions.includes('view gamechanger')
+			) {
 				const { extSearchFields = [], extRetrieveFields = [] } = this.constants.EDA_ELASTIC_SEARCH_OPTS;
 				body.extSearchFields = extSearchFields.map((field) => field.toLowerCase());
 				body.extStoredFields = extRetrieveFields.map((field) => field.toLowerCase());
@@ -197,13 +202,13 @@ class EdaSearchHandler extends SearchHandler {
 	async queryContractMods(req, userId) {
 		try {
 			const clientObj = { esClientName: 'eda', esIndex: this.constants.EDA_ELASTIC_SEARCH_OPTS.index };
-			const permissions = req.permissions ? req.permissions : [];
+			const permissions = req.permissions ? req.permissions.map((permission) => permission.toLowerCase()) : [];
 			const { esClientName, esIndex } = clientObj;
 			const { awardID, isSearch } = req.body;
 			const { id, idv } = this.edaSearchUtility.splitAwardID(awardID);
 
 			let esQuery = '';
-			if (permissions.includes('View EDA') || permissions.includes('eda Admin')) {
+			if (permissions.includes('view eda') || permissions.includes('eda admin')) {
 				esQuery = this.edaSearchUtility.getEDAContractQuery(id, idv, false, isSearch, userId);
 			} else {
 				throw 'Unauthorized';
@@ -264,14 +269,14 @@ class EdaSearchHandler extends SearchHandler {
 	async queryBaseAwardContract(req, userId) {
 		try {
 			const clientObj = { esClientName: 'eda', esIndex: this.constants.EDA_ELASTIC_SEARCH_OPTS.index };
-			const permissions = req.permissions ? req.permissions : [];
+			const permissions = req.permissions ? req.permissions.map((permission) => permission.toLowerCase()) : [];
 			const { esClientName, esIndex } = clientObj;
 			const { awardID } = req.body;
 
 			const { id, idv } = this.edaSearchUtility.splitAwardID(awardID);
 
 			let esQuery = '';
-			if (permissions.includes('View EDA') || permissions.includes('eda Admin')) {
+			if (permissions.includes('view eda') || permissions.includes('eda admin')) {
 				esQuery = this.edaSearchUtility.getEDAContractQuery(id, idv, true, false, userId);
 			} else {
 				throw 'Unauthorized';
@@ -312,13 +317,13 @@ class EdaSearchHandler extends SearchHandler {
 	async querySimilarDocs(req, userId) {
 		try {
 			const clientObj = { esClientName: 'eda', esIndex: this.constants.EDA_ELASTIC_SEARCH_OPTS.index };
-			const permissions = req.permissions ? req.permissions : [];
+			const permissions = req.permissions ? req.permissions.map((permission) => permission.toLowerCase()) : [];
 			const { esClientName, esIndex } = clientObj;
 			const { body } = req;
 			const { issueOfficeDoDAAC, issueOfficeName } = body;
 
 			let esQuery = '';
-			if (permissions.includes('View EDA') || permissions.includes('eda Admin')) {
+			if (permissions.includes('view EDA') || permissions.includes('eda admin')) {
 				esQuery = this.edaSearchUtility.getElasticsearchPagesQuery(
 					{ ...body, limit: 5, edaSearchSettings: { issueOfficeDoDAAC, issueOfficeName } },
 					userId
@@ -360,8 +365,8 @@ class EdaSearchHandler extends SearchHandler {
 		const { functionName } = req.body;
 
 		try {
-			const permissions = req.permissions ? req.permissions : [];
-			if (permissions.includes('View EDA') || permissions.includes('eda Admin')) {
+			const permissions = req.permissions ? req.permissions.map((permission) => permission.toLowerCase()) : [];
+			if (permissions.includes('view EDA') || permissions.includes('eda admin')) {
 				switch (functionName) {
 					case 'queryContractMods':
 						return await this.queryContractMods(req, userId);
