+++ frontend/src/components/modules/jbook/jbookContext.js
@@ -397,6 +397,7 @@ const initState = {
 
 	// jbook portfolios
 	portfolios: [],
+	selectedPortfolio: 'General',
 };
 
 const init = (initialState) => {
